Hotfix/python and roles (#25)
* fix: python now run in docker and start adding roles limitation

* fix: add field to not delete picture profile if not set in update

* feat: Handle 403 Forbidden errors during GitHub repository creation

* fix: add protection to endpoints

+++ src/user/user.service.ts
@@ -317,20 +317,26 @@ export class UserService {
     }
 
     if (
+      updateUserDto.profilePicturePath !== undefined &&
       user.profilePicturePath &&
       updateUserDto.profilePicturePath !== user.profilePicturePath
     ) {
       await this.S3BucketService.deleteFile("users", user.profilePicturePath);
     }
 
+    const dataToUpdate: any = {
+      ...updateUserDto,
+      github: updateUserDto.github ?? undefined,
+      discord: updateUserDto.discord ?? undefined,
+    };
+
+    if (updateUserDto.profilePicturePath !== undefined) {
+      dataToUpdate.profilePicturePath = updateUserDto.profilePicturePath;
+    }
+
     return await this.prisma.user.update({
       where: { id },
-      data: {
-        ...updateUserDto,
-        profilePicturePath: updateUserDto.profilePicturePath ?? null,
-        github: updateUserDto.github ?? undefined,
-        discord: updateUserDto.discord ?? undefined,
-      },
+      data: dataToUpdate,
     });
   }
